# schemas/platform/C-010.openapi.json
{
  "components": {
    "schemas": {
...
      "AssetResolutionRequest": {
        "additionalProperties": false,
        "properties": {
          "allow_inactive": {
            "default": false,
            "title": "Allow Inactive",
            "type": "boolean"
          },
          "asset_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetType"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "input_namespace": {
            "maxLength": 128,
            "title": "Input Namespace",
            "type": "string"
          },
          "input_value": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Input Value",
            "type": "string"
          }
        },
        "required": [
          "input_namespace",
          "input_value"
        ],
        "title": "AssetResolutionRequest",
        "type": "object"
      },
...
      "AssetType": {
        "enum": [
          "stock",
          "index",
          "etf",
          "convertible_bond",
          "fund",
          "future",
          "fx",
          "commodity"
        ],
        "title": "AssetType",
        "type": "string"
      },
...
      "PlatformAPIError": {
        "additionalProperties": false,
        "description": "Stable public error fields required by C-010.",
        "properties": {
          "code": {
            "pattern": "^[A-Z][A-Z0-9_]{2,63}$",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "pattern": "^req_[0-9a-f]{32}$",
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "retryable",
          "request_id"
        ],
        "title": "PlatformAPIError",
        "type": "object"
      },
      "PlatformErrorEnvelope": {
        "additionalProperties": false,
        "description": "C-010 error response.",
        "examples": [
          {
            "error": {
              "code": "RESOURCE_NOT_FOUND",
              "details": {},
              "message": "资源不存在",
              "request_id": "req_0123456789abcdef0123456789abcdef",
              "retryable": false
            }
          }
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/PlatformAPIError"
          }
        },
        "required": [
          "error"
        ],
        "title": "PlatformErrorEnvelope",
        "type": "object"
      },
      "PlatformListEnvelope": {
        "additionalProperties": false,
        "description": "C-010 list success response with opaque cursor metadata.",
        "examples": [
          {
            "data": [
              {
                "resource_id": "example"
              }
            ],
            "meta": {
              "data_snapshot_id": null,
              "generated_at": "2026-08-27T12:00:00+08:00",
              "request_id": "req_0123456789abcdef0123456789abcdef",
              "schema_version": "1.0.0",
              "warnings": []
            },
            "page": {
              "cursor": null,
              "has_more": false,
              "limit": 50,
              "next_cursor": null
            }
          }
        ],
        "properties": {
          "data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PlatformResponseMeta"
          },
          "page": {
            "$ref": "#/components/schemas/PlatformPage"
          }
        },
        "required": [
          "data",
          "meta",
          "page"
        ],
        "title": "PlatformListEnvelope",
        "type": "object"
      },
      "PlatformPage": {
        "additionalProperties": false,
        "description": "Opaque cursor metadata for C-010 list responses.",
        "properties": {
          "cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "limit": {
            "maximum": 1000,
            "minimum": 1,
            "title": "Limit",
            "type": "integer"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "cursor",
          "next_cursor",
          "limit",
          "has_more"
        ],
        "title": "PlatformPage",
        "type": "object"
      },
      "PlatformResponseMeta": {
        "additionalProperties": false,
        "description": "C-010 metadata shared by all successful platform responses.",
        "properties": {
          "data_snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Snapshot Id"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "request_id": {
            "pattern": "^req_[0-9a-f]{32}$",
            "title": "Request Id",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "schema_version",
          "generated_at",
          "data_snapshot_id",
          "warnings"
        ],
        "title": "PlatformResponseMeta",
        "type": "object"
      },
      "PlatformSuccessEnvelope": {
        "additionalProperties": false,
        "description": "C-010 object success response.",
        "examples": [
          {
            "data": {
              "resource_id": "example"
            },
            "meta": {
              "data_snapshot_id": null,
              "generated_at": "2026-08-27T12:00:00+08:00",
              "request_id": "req_0123456789abcdef0123456789abcdef",
              "schema_version": "1.0.0",
              "warnings": []
            }
          }
        ],
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "meta": {
            "$ref": "#/components/schemas/PlatformResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "PlatformSuccessEnvelope",
        "type": "object"
      },
      "PriorityClass": {
        "enum": [
          "P0_DATA_CERTIFICATION",
          "P1_FORMAL_SIGNAL",
          "P2_MARKET_REVIEW",
          "P3_USER_INTERACTIVE",
          "P4_RESEARCH",
          "P5_PREVIEW_AND_MAINTENANCE"
        ],
        "title": "PriorityClass",
        "type": "string"
      },
...
      "ResourceRef": {
        "additionalProperties": false,
        "properties": {
          "resource_id": {
            "title": "Resource Id",
            "type": "string"
          },
          "resource_type": {
            "$ref": "#/components/schemas/ResourceType"
          }
        },
        "required": [
          "resource_type",
          "resource_id"
        ],
        "title": "ResourceRef",
        "type": "object"
      },
      "ResourceType": {
        "enum": [
          "task",
          "attempt",
          "data_snapshot",
          "feature_snapshot",
          "observation_snapshot",
          "fact_pack",
          "research",
          "review",
          "strategy",
          "backtest_run",
          "prediction",
          "artifact",
          "report",
          "provider_run",
          "sector",
          "taxonomy",
          "indicator",
          "raw_object",
          "canonical_partition",
          "feature_partition",
          "fact_block",
          "claim",
          "watch_condition",
          "quality_report",
          "request",
          "checkpoint",
          "backup",
          "deployment"
        ],
        "title": "ResourceType",
        "type": "string"
      },
...
      "TaskCancelRequest": {
        "additionalProperties": false,
        "properties": {
          "reason_code": {
            "default": "TASK_CANCEL_REQUESTED",
            "title": "Reason Code",
            "type": "string"
          }
        },
        "title": "TaskCancelRequest",
        "type": "object"
      },
...
      "TaskCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "created_from_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Created From Task Id"
          },
          "force": {
            "default": false,
            "title": "Force",
            "type": "boolean"
          },
          "force_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Force Reason"
          },
          "input_refs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ResourceRef"
            },
            "title": "Input Refs",
            "type": "array"
          },
          "max_attempts": {
            "default": 3,
            "maximum": 100,
            "minimum": 1,
            "title": "Max Attempts",
            "type": "integer"
          },
          "priority_class": {
            "$ref": "#/components/schemas/PriorityClass",
            "default": "P5_PREVIEW_AND_MAINTENANCE"
          },
          "priority_value": {
            "default": 100,
            "maximum": 2147483647,
            "minimum": 0,
            "title": "Priority Value",
            "type": "integer"
          },
          "request_source": {
            "default": "platform_api",
            "maxLength": 64,
            "minLength": 1,
            "title": "Request Source",
            "type": "string"
          },
          "requested_by": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Requested By",
            "type": "string"
          },
          "requirements": {
            "additionalProperties": true,
            "title": "Requirements",
            "type": "object"
          },
          "task_schema_version": {
            "default": "1.0.0",
            "maxLength": 32,
            "pattern": "^[0-9]+\\.[0-9]+\\.[0-9]+$",
            "title": "Task Schema Version",
            "type": "string"
          },
          "task_type": {
            "const": "artifact_orphan_dry_run",
            "title": "Task Type",
            "type": "string"
          }
        },
        "required": [
          "task_type",
          "requested_by"
        ],
        "title": "TaskCreateRequest",
        "type": "object"
      },
...
      "TaskRetryRequest": {
        "additionalProperties": false,
        "properties": {
          "reason_code": {
            "default": "TASK_RETRY_REQUESTED",
            "title": "Reason Code",
            "type": "string"
          }
        },
        "title": "TaskRetryRequest",
        "type": "object"
      },
      "TaskState": {
        "enum": [
          "ACCEPTED",
          "QUEUED",
          "BLOCKED",
          "LEASED",
          "RUNNING",
          "RETRY_WAIT",
          "SUCCEEDED",
          "DEGRADED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "TaskState",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "C-002/C-003/C-007/C-010/C-011 public platform components. Only explicitly implemented runtime resources appear under /api/platform/v1; Legacy /api/v1 responses are not wrapped.",
    "title": "Visory Platform API Contracts",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/platform/v1/asset-resolutions": {
      "post": {
        "operationId": "resolveAssetIdentity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetResolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformSuccessEnvelope"
                }
              }
            },
            "description": "C-002 resolution result wrapped by C-010"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformErrorEnvelope"
                }
              }
            },
            "description": "Invalid request"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformErrorEnvelope"
                }
              }
            },
            "description": "Identity repository unavailable"
          }
        },
        "summary": "Resolve an asset identity without guessing"
      }
    },
...
    "/api/platform/v1/tasks": {
      "get": {
        "operationId": "listTasks",
        "parameters": [
          {
            "in": "query",
            "name": "tab",
            "schema": {
              "enum": [
                "active",
                "blocked",
                "failed",
                "history"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "task_state",
            "schema": {
              "$ref": "#/components/schemas/TaskState"
            }
          },
          {
            "in": "query",
            "name": "task_type",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "priority_class",
            "schema": {
              "$ref": "#/components/schemas/PriorityClass"
            }
          },
          {
            "in": "query",
            "name": "requested_by",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "created_from",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "created_to",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resource_id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformListEnvelope"
                }
              }
            },
            "description": "Task list"
          }
        },
        "summary": "List durable platform tasks"
      },
      "post": {
        "operationId": "createTask",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformSuccessEnvelope"
                }
              }
            },
            "description": "Task created or idempotently replayed"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid idempotency key"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformErrorEnvelope"
                }
              }
            },
            "description": "Idempotency conflict"
          }
        },
        "summary": "Create a durable platform task"
      }
    },
    "/api/platform/v1/tasks/events": {
      "get": {
        "operationId": "streamTaskEvents",
        "parameters": [
          {
            "in": "query",
            "name": "after_event_id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Resumable task event stream"
          }
        },
        "summary": "Stream task state events"
      }
    },
    "/api/platform/v1/tasks/{task_id}": {
      "get": {
        "operationId": "getTask",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformSuccessEnvelope"
                }
              }
            },
            "description": "Task details"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformErrorEnvelope"
                }
              }
            },
            "description": "Task not found"
          }
        },
        "summary": "Get a durable platform task"
      }
    },
    "/api/platform/v1/tasks/{task_id}/cancellations": {
      "post": {
        "operationId": "requestTaskCancellation",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformSuccessEnvelope"
                }
              }
            },
            "description": "Cancellation requested"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformErrorEnvelope"
                }
              }
            },
            "description": "Cancellation rejected"
          }
        },
        "summary": "Request task cancellation"
      }
    },
    "/api/platform/v1/tasks/{task_id}/events": {
      "get": {
        "operationId": "streamTaskEventsForTask",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_event_id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Resumable task event stream"
          }
        },
        "summary": "Stream state events for one task"
      }
    },
    "/api/platform/v1/tasks/{task_id}/retries": {
      "post": {
        "operationId": "requestTaskRetry",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRetryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformSuccessEnvelope"
                }
              }
            },
            "description": "Retry requested"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformErrorEnvelope"
                }
              }
            },
            "description": "Retry rejected"
          }
        },
        "summary": "Request task retry"
      }
    }
  }
}